requirements.txt
python-hcl2==5.1.1
Jinja2==3.1.4
